Improved clone

+++ src/net/psj/Particles/ParticleClone.java
@@ -18,18 +18,17 @@ public class ParticleClone extends Particle {
 	public boolean update() {
 		boolean ret = super.update();
 		if (ctype != -0) {
-			for (int x2 = (int) x - 1; x2 < x + 1; x2++)
-				for (int y2 = (int) y - 1; y2 < y + 1; y2++)
-					PowderSimJ.ptypes.create_part(x2, y2, ctype, false);
+			for (float x2 = x - 1; x2 < x + 1; x2++)
+				for (float y2 = y - 1; y2 < y + 1; y2++)
+					PowderSimJ.ptypes.create_part((int)x2, (int)y2, ctype, false);
 		} else {
-			int rx, ry;
+			float rx, ry;
 			for (rx = -1; rx < 2; rx++)
 				for (ry = -1; ry < 2; ry++)
 					if (x + rx >= 0 && y + ry > 0 && x + rx < PowderSimJ.width
 							&& y + ry < PowderSimJ.height
 							&& (rx != 0 || ry != 0)) {
-						Particle r = ParticleData.parts[ParticleData.pmap[(int) y
-								+ ry][(int) x + rx]];
+						Particle r = ParticleData.parts[ParticleData.pmap[(int) y + (int) ry][(int) x + (int) rx]];
 						if (r == null)
 							continue;
 						if (r.id == 0)
